boosts now apply the attribute changes, also other functionality fixes
Boosts now apply the necessary attribute upgrades, and remove the attribute upgrades when unselected.  Also functionality fixes for when a boost which makes an attribute meet the requirement for an ability is removed (so for example: if an ability is selected and one of its requirements is met because of a boost that is applied, and if that boost is then removed after the ability becomes invalid and requires the boost to be applied again in order for it to be valid).